Clean up AnyClient type
